improve fix to Date::Manip deprecated TZ warnings
Use DateManipVersion() rather than $Date::Manip::VERSION to determine
which version of Date::Manip is in effect.  This is needed because 6.x
versions of Date::Manip will load/use Date::Manip::DM5 and use that on
older perls (5.8.x).  In this case $Date::Manip::VERSION is 6.x, but
DateManipVersion() is 5.x.